+++ data/src/main/java/com/kfouri/cryptoprice2/data/network/model/CurrencyNetworkEntity.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class CurrencyNetworkEntity (
     @SerializedName("name") val name: String,
-    @SerializedName("price") val price: Float,
+    @SerializedName("price") val price: Double,
     @SerializedName("icon") val icon: String,
 )
